requirements2.txt
certifi==2024.8.30
charset-normalizer==3.3.2
greenlet==3.1.1
idna==3.10
numpy==1.26.4
packaging==24.1
pandas==1.5.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==1.4.53
sqlalchemy-redshift==0.8.14
urllib3==2.2.3
pytest==8.3.2
apache-airflow==2.10.1
